+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Variables created with `verb-var` are now buffer-local, and are not defined as global values.
 - Code tags are now evaluated with the current `.org` buffer as the current buffer.
 - Added `verb-headers-get` function.
+- Regular expressions can now be used as keys in `verb-content-type-handlers`, to easily assign one handler to one or more content types.
 
 ## **2.6.0** - 2020-02-03
 - Removed `verb-max-redirections` variable, as it wasn't working as intended. Use `url-max-redirections` instead.